[HWKMETRICS-495] upgrade to cassalog 0.4.0 and add verification checks
The verification checks are used in restart scenarios to help determine whether
or not a change, such as creating a table, has already been applied. Applying
the schema change and updating the changelog table cannot be done atomically
so we need to do some extra work to make sure we are in a known, good state.