[HWKMETRICS-524] add flag for disabling compression job
The job can be enabled/disabled with either the system property

  -Dhawkular.metrics.jobs.compression.enabled=false

or using the environment variable

  export COMPRESSION_JOB_ENABLED=false

The job actually still runs but is just a no-op. This is important from a job
scheduling standpoint. Suppose the job is disabled just before 13:00, and it is
re-enabled at 23:00. Raw data ingested between 11:00 and 21:00 will not get
compressed.