[HWKMETRICS-564] fix several bugs
When jobs fall behind and are rescheduled for immediate execution, we first
renew the job lock. The job should only be executed again if the lock is
successfully renewed. We were always executing the job. This is fixed.

When there are jobs to execute across multiple time slices, we were executing
things out of order. This is fixed simply by replacing flatMap with concatMap
in a few places.

When retrying a job execution, we were updating finished_jobs_idx with the
wrong timestamp. We were using the trigger time instead of the time in which
the job is scheduled. In this scenario, the scheduled time != the trigger time.