[HWKMETRICS-583] some small changes to job locks
The job lock name now includes the hostname and the queries for acquiring and
renewing locks have been modified so that a job scheduler instance can
reacquire a lock it already holds.

This commit also removes the call to reschedule a job immediately when it falls
behind schedule. I removed it because it is 1) it an optimization that is not
necessary and 2) it was causing some inconsistent behavior due to a race
condition when the indexes are read and updated.