[HWKMETRICS-556] use TWCS for data_compressed table
We are also switching back to STCS for the data table because we are now doing
range deletes as part of the compression job. Eventually I think we'll switch
over to using what are essentially temporary tables for raw data that we just
drop once we are done writing to them and have compressed the data.

[HWKEMTRICS-556] license header

[HWKMETRICS-556] this is what i get for trying to catch up with work on the weekend

[HWKMETRICS-556] cehckstyle