[HWKMETRICS-660] use logged batches for adding tags
This commit updates the methods for adding tags which is used by endpoints like

PUT /hawkular/metrics/gauges/{id}/tags

I did not update the method for creating metrics since it is called by
Heapster. The create metric operation also has a flag that will cause it to
use a light weight transaction. LWTs cannot be batched AFAIK so we will have
think more about how we want to handle this, in a separate ticket though. I
did also update the deleteMetric method as well to keep tags consistent there.