[HWKMETRICS-763] fix bug in scheduler error handling
I found the bug today which is in SchedulerImpl.java. At some point a while back
I added retry support for when jobs fails. You provide a RetryPolicy that
determines if and when a job should be retried. This was primarily intended for
non-repeating, single-execution jobs. For repeating jobs, i.e., the compression
job, the job is supposed to be executed according to its trigger. Suppose the
compression job is scheduled to run at 15:00 and repeats every two hours. If
the 15:00 execution fails, the scheduler is supposed to retry the job (for all
intensive purposes) immediately with the same trigger time of 15:00. When the
job completes normally, its trigger will get updated in the database to the
next execution time of 17:00. If it already later than 17:00, then the
scheduler will again execute the job right away.

There was a bug with the error handling such that the job wasn't getting
retried for the 15:00 execution. The trigger was getting advanced and set to
17:00. The temp table being compressed gets dropped at the end of the
compression job. Unless the failure was dropping the table, we end up with
orphaned temp tables any time the job fails.

This commit does away with the RetryPolicy since it is not used. This
means that when a job fails, regardless of whether or not it is
repeating, it will be retried with the same trigger until it completes
normally.