[HWKMETRICS-763] add "local" job to check for expired temp tables
Due to the bug in SchedulerImpl.java which was fixed in my previous
commit, there are openshift clusters with literally hundreds of expired
temp tables. They are expired in the sense that they are older than the
data retention and therefore do not contain any live data.

This commit adds a "local" job that checks for and drops expired tables.
The job is local in the sense that it is local to the hawkular-metrics
server. It is not run via the job scheduler.