[HWKMETRICS-787] optimize for pod_id queries (#961)
* Optimize Openshift tag queries

* [HWKMETRICS-787] clean up pod_id tag query logic and add tests

* [HWKMETRICS-787] clean up pod_id query and add some debug logging

I removed the reduce operation since it is not needed. This whole
optimization is for a particular tag key. There can only be one instance
of a tag key per request; hence, the reduce operation is not needed.

* [HWKMETRICS-787] disable failing test

I spent a few minutes investigating, but I am not going to wast any more
time. The delete tenant functionality has never been used, and the REST
integration tests are tricky due to using the "virtual clock" which in
turn makes these tests brittle.

I will open a separate ticket for disabling the delete tenant
functionality. We can leave the endpoints intact but return an
appropriate status and error message indicating that it is no longer
supported.