[HWKMETRICS-797] wait for schema agreement before preparing queries (#989)
* [HWKMETRICS-797] wait for schema agreement before preparing queries

* [HWKMETRICS-797] query system.peers on all nodes

Do a token range scan to query system.peers on all nodes so that we can
determine the schema versions that every node knows about.

* [HWKMETRICS-797] replace schema version token scan with simple retry logic

The previous commit introduced a token range scan over the system.peers
table in order to determine when there is schema convergence. That
approach is complicated and does not cover some corner cases. It is
simpler to retry preparing statements after some delay until it
succeeds.

* [HWKMETRICS-797] adjust logging based on review